test: DCI 패턴을 적용한 CouponStatusTest 적용
